Merged PR 12231: bug fix in factors for {word} class token
This fixes a bug in how the factor types for the `{word}` lemmas are determined. The `{word}` lemma is used to represent phrase fixes (when using the class-based phrasefix mechanism).

The bug was in a routine that determines the set of factor types for a given lemma. In order to be general across all configuration options, this routine is based on examples, for which it determines the factors given the current configuration.

The bug was that this routine applied sentence-piece breaking to the example words. In some non-Latin languages, there are so few latin words in the training that SentencePiece breaks those words into individual letters. If we also have the `DistinguishWordInitialPieces` option enabled, a _ gets prepended, and SentencePiece may split it into a separate token, for example "Hello" --> "_ H E L L O". The routine reads the factor types of the example word from the first token. In this case, that is wrong, as the first token is _ and therefore does not have the same factor set as the actual example word.

The fix is to not apply SentencePiece breaking to example words. I also sanity added a check that the example words are not broken into sub-tokens.

Related work items: #107678