Merged PR 12304: prototype of libsentencepiece wrapper in C
This is a prototype of a wrapper of libsentencepiece in C, per discussion with Fai.

There is no proper makefile yet. The CPP file contains the `clang` command line to build; run `a.out` to test it.

The sentence-piece library must be installed from source to get the header and the lib. The CPP file contains a pointer to the installation instructions.